express 서브

+++ meet/src/main/java/hanseimeet/meet/controller/user/UserController.java
@@ -44,13 +44,33 @@ public class UserController {
 
     @PutMapping("/user")
     public void updateUser(@RequestBody UserUpdateRequest request){
+        String readSql = "SELECT * FROM user WHERE id = ?";
+        // 하나라도 있다면 0
+        boolean isUserNotExist = jdbcTemplate.query(readSql,(rs, rowNum) -> 0, request.getId()).isEmpty();
+        
+        if(isUserNotExist){
+            throw new IllegalArgumentException();
+        }
         String sql = "UPDATE user SET name = ?  WHERE id = ?";
         jdbcTemplate.update(sql, request.getName(),  request.getId());
     }
 
     @DeleteMapping("/user")
     public void deleteUser(@RequestParam String name, String major){
+        String deleteSql = "SELECT * FROM user WHERE name = ?";
+        // 하나라도 있다면 0
+        boolean isUserNotExist = jdbcTemplate.query(deleteSql,(rs, rowNum) -> 0, name).isEmpty();
+
+        if(isUserNotExist){
+            throw new IllegalArgumentException();
+        }
         String sql = "DELETE FROM user WHERE name = ? OR major = ?";
         jdbcTemplate.update(sql, name, major);
+    }
+
+    // 에러 던지기
+    @GetMapping("/user/error-test")
+    public void errTest(){
+        throw new IllegalArgumentException();
     }
 }
